[UPDATE] move token dto
